refactor: ts -> tsx로 변경
